feat: add runLanguageClient
