+++ packages/examples/src/common/language-server-runner.ts
@@ -9,7 +9,7 @@ import express from 'express';
 import { getLocalDirectory } from '../utils/fs-utils.js';
 import { upgradeWsServer } from './server-commons.js';
 import { LanguageServerRunConfig } from './model.js';
-
+/** LSP server runner */
 export const runLanguageServer = (
     languageServerRunConfig: LanguageServerRunConfig
 ) => {
